pyproject.toml
[tool.poetry]
name = "ffmpeg-testing"
version = "0.1.0"
description = ""
authors = ["Delphine Tai-Beauchamp <delphinetaibeauchamp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.114.0"
uvicorn = "^0.30.6"
ffmpeg-python = "^0.2.0"
python-multipart = "^0.0.9"
loguru = "^0.7.2"
pytest = "^8.3.3"
httpx = "^0.27.2"
openai = "^1.40.0"
langchain-openai = "^0.2.0"
pydantic = "^2.9.1"
python-dotenv = "^1.0.1"
streamlit = "^1.38.0"
pytube = "^15.0.0"
youtube-transcript-api = "^0.6.2"
migrate = "^0.3.8"
numpy = "^2.1.1"
moviepy = "^1.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
